requirements.txt
web3==6.15.1
prometheus_client==0.19.0
aiohttp==3.9.1
pyyaml==6.0.1